Refactor code to 2 packages
