+++ Cago.toml
@@ -1,0 +1,3 @@
+
+[workspace]
+members = ["packages/*"]
