+++ packages/std/src/lib.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+mod macros;
+
+pub mod ext;
+pub mod oei;
