+++ packages/vm/Cargo.toml
@@ -1,18 +1,16 @@
 [package]
-name = "owasm"
-version = "0.1.11"
+name = "owasm-vm"
+version = "0.1.12"
 authors = ["Band Protocol <connect@bandprotocol.com>"]
 edition = "2018"
 license = "Apache-2.0"
-repository = "https://github.com/bandprotocol/owasm"
-description = "Standard library for oracle scripts on BandChain"
+repository = "https://github.com/bandprotocol/owasm/tree/master/packages/vm"
+description = "Oracle script VM runtime"
 readme = "README.md"
 
 [dependencies]
 assert_matches = "1.3.0"
 hex = "0.4"
-json = "0.12.0"
-num = "0.2.1"
 parity-wasm = "0.41"
 pwasm-utils = "0.12"
 wasmer = { version = "1.0.2", default-features = false, features = ["jit", "singlepass", "compiler"] }
