+++ packages/vm/README.md
@@ -1,0 +1,9 @@
+# Owasm
+
+[![owasm on crates.io](https://img.shields.io/crates/v/owasm.svg)](https://crates.io/crates/owasm)
+
+The standard libraries for building BandChain oracle scripts. Code in this package is split into 3 packages -- `core`, `ext`, and `oei`.
+
+`core` - a library for building VM runtime
+`ext` - a library is helper package for writing the oracle scripts 
+`oei` - a library containing functions for querying data from BandChain
