+++ packages/vm/src/error.rs
